requirements.txt
beautifulsoup4==4.10.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
jsonfield==3.1.0
mouse==0.7.1
pillow==8.4.0
pyopenssl==21.0.0
stripe==2.63.0
virtualenv==20.8.1
gunicorn
pytz
Faker